Consolidate Buffer tests
